Refactor: keystore 패키지 추가 및 관련 파일 이동
