+++ core/security/src/main/java/com/threegap/bitnagil/security/keystore/KeyProvider.kt
@@ -1,4 +1,4 @@
-package com.threegap.bitnagil.security
+package com.threegap.bitnagil.security.keystore
 
 import javax.crypto.SecretKey
